+++ src/components/tooltip/tooltip.js
@@ -64,6 +64,7 @@ const TOOLTIP_DEFAULTS = {
   tooltipElement: null,
   parentElement: null,
   keepOpen: false,
+  isRangeDatepicker: false,
   extraClass: null,
   placementOpts: {},
   maxWidth: null,
@@ -937,7 +938,11 @@ Tooltip.prototype = {
    */
   position() {
     this.setTargetContainer();
-    this.tooltip[0].classList.remove('is-hidden');
+
+    // Popup is range datepicker and should not be shown until range is selected
+    if (!this.settings.isRangeDatepicker) {
+      this.tooltip[0].classList.remove('is-hidden');
+    }
 
     const self = this;
     const distance = this.isPopover ? 20 : 10;